Added constants and RLE of line info

+++ src/include/memory.h
@@ -7,11 +7,11 @@
         ((capacity) < 8 ? 8 : capacity * 2)
 
 #define GROW_ARRAY(type, pointer, oldCount, newCount) \
-        (type*)reallocate(pointer, sizeof(type) * (oldCount), \
-        sizeof(type) * (newCount))
+        (type*)reallocate(pointer, sizeof(type) * (size_t)(oldCount), \
+        sizeof(type) * (size_t)(newCount))
 
 #define FREE_ARRAY(type, pointer, oldCount) \
-        reallocate(pointer, sizeof(type) * (oldCount), 0)
+        reallocate(pointer, sizeof(type) * (size_t)(oldCount), 0)
 
 void* reallocate(void* pointer, size_t oldSize, size_t newSize);
